added encryption detection to keyfiles

+++ qd/cae/dyna_cpp/utility/FileUtility.hpp
@@ -7,15 +7,23 @@
 
 namespace qd {
 
-bool check_ExistanceAndAccess(std::string);
-
-std::vector<std::string> globVector(std::string);
-
-std::vector<std::string>
-read_textFile(std::string _filepath);
+bool
+check_ExistanceAndAccess(const std::string&);
 
 std::vector<std::string>
-findDynaResultFiles(std::string _base_file);
+glob_vector(const std::string&);
+
+std::vector<std::string>
+read_text_file(const std::string& _filepath);
+
+std::vector<char>
+read_binary_file(const std::string& _filepath);
+
+std::vector<std::string>
+find_dyna_result_files(const std::string& _base_file);
+
+double
+get_entropy(const std::vector<char>& _buffer);
 
 void
 delete_file(const std::string& _path);
